+++ R/order_by.R
@@ -166,13 +166,9 @@ to_sql.relop_orderby <- function (x,
     ot <- c(ot, rev_ot)
   }
   subcols <- calc_used_relop_orderby(x, using=using)
-  qlimit = limit
-  if(!getDBOption(db, "use_pass_limit", TRUE)) {
-    qlimit = NULL
-  }
   subsql_list <- to_sql(x$source[[1]],
                         db = db,
-                        limit = qlimit,
+                        limit = NULL, # can't pass down limit from order_by
                         source_limit = source_limit,
                         indent_level = indent_level + 1,
                         tnum = tnum,